refactor(cat): split cat-editor-panel into isolated sub-components
Break the 760-line monolithic CatEditorPanel into focused modules for
better API boundaries and code-splitting. Lazy-load below-the-fold
sections (AI recommendation, format checks, comments) via React.lazy.
